requirements.txt
git+https://github.com/openai/whisper.git
sounddevice
numpy
scipy
python-docx
torch
